+++ src/features/spaceOwner/CreateRoom/style.css.ts
@@ -117,32 +117,6 @@ export const businessDayText = style({
   color: theme.white,
 });
 
-export const businessTimeContainer = style({
-  width: '100%',
-  ...flex.COLUMN_VERTICAL,
-  gap: '20px',
-});
-
-export const startTimeWrapper = style({
-  width: '100%',
-  ...flex.BETWEEN,
-});
-
-export const startTimeLabel = style({
-  ...font.p1,
-  color: theme.white,
-});
-
-export const timeWrapper = style({
-  ...flex.FLEX,
-  gap: '12px',
-});
-
-export const time = style({
-  ...font.p1,
-  color: theme.white,
-});
-
 export const roomWrapper = style({
   width: '100%',
   ...flex.COLUMN_FLEX,
